requirements.txt
fastapi
uvicorn
pytesseract
Pillow
PyMuPDF
sqlalchemy
pydantic
streamlit
requests
